Add files via upload
This is the classification of movement file including the feedback feature.
The feedback will be as following: The proctor will be holding a mobile device it's camera will be connected to the laptop that the child is sitting in front of , it will classify it's movements based on the identified/classified ones. Thus it will show either negative or positive feedback.